fixed posts site and navigation between them, and cleaned up with components and made it more structered
