+++ .expo/packager-info.json
@@ -1,0 +1,4 @@
+{
+  "expoServerPort": 19000,
+  "packagerPort": 19000
+}
