+++ app/Posts/Posts.style.js
@@ -1,0 +1,3 @@
+/** @format */
+
+import { StyleSheet } from "react-native";
