+++ app/Posts/posts.js
@@ -1,0 +1,10 @@
+/** @format */
+
+import React from "react";
+import { Text } from "react-native";
+
+const Posts = () => {
+  return <Text>Her skal alle postene ligge</Text>;
+};
+
+export default Posts;
