+++ app/_layout.js
@@ -4,7 +4,13 @@ import { Stack } from "expo-router";
 import * as React from "react";
 
 const Layout = () => {
-  return <Stack />;
+  return (
+    <Stack
+      screenOptions={{
+        headerShown: false,
+      }}
+    ></Stack>
+  );
 };
 
 export default Layout;
