--- app/video.js
@@ -1,14 +1,0 @@
-/** @format */
-
-import { View, Text } from "react-native";
-import * as React from "react";
-
-const Video = () => {
-  return (
-    <View>
-      <Text>Hello World Video</Text>
-    </View>
-  );
-};
-
-export default Video;
