+++ package.json
@@ -9,11 +9,13 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "expo": "^52.0.16",
-    "expo-constants": "~17.0.2",
-    "expo-linking": "~7.0.2",
+    "@react-navigation/material-bottom-tabs": "^6.2.29",
+    "@react-navigation/native": "^6.10.1",
+    "expo": "~52.0.28",
+    "expo-constants": "~17.0.5",
+    "expo-linking": "~7.0.5",
     "expo-router": "~4.0.2",
-    "expo-splash-screen": "~0.29.7",
+    "expo-splash-screen": "^0.29.21",
     "expo-status-bar": "~2.0.0",
     "react": "18.3.1",
     "react-dom": "18.3.1",
